Move upload data models
